backend/requirements_api.txt
fastapi
uvicorn
prefect
google-cloud-bigquery
google-cloud-storage
vaderSentiment
python-multipart
clearml